Fix getting source keyword in a scan for VEX2
The source parameter changed to a multi-field parameter which broke
retrieving the keyword. The source parameter is one of the few parameters
that have a dedicated retrieval call. It will be necessary to add a
call to get an arbitrary source field for VEX2. The old form should be
preserved for backward compatibility.

Thanks to Harro for reporting this wasn't working for VEX2.