Add retrieving VEX version for parsed files
Added get_vex_rev and fget_vex_rev.

Although switching between reading and writing VEX files is possible,
the current implementation will not support interrupting the writing a
VEX file (using the "put" functions) with reading a VEX file. The
latter will probably work if the VEX version of the file being read is
the same as the file being written, but it should still be avoided
(the first operation in building up a VEX file for writing should be
setting the version number). This has true since commit
bd1b08288019e73d90f343526ba769a1020ae62c. This is caused by the fact
that reading and writing must share the internal data structure,
vex_version.

It is unfortunate there is more than one copy of the version string,
but that is necessary if more than one file is operated on. One is
needed for each file. One is used internally by the parser/writer to
keep things simple.