'struct vector` clashes with C++ std::vector
When using the vex parser in C++ code using e.g. std::vector the 'struct
vector` declaration/definition will clash with that one.

As the type is strictly an internal type to the vex parser library, the
struct can safely be renamed to something that doesn't clash with C++'s
stad::vector.