Add support for $THREADS block
At JIVE the $THREADS block has been (unofficially) used to describe VDIF
data streams for a long time. The block served as basis for the VEX2
$DATASTREAMS block and contains basically the same information.

Because JIVE has ~a decade's worth of VEX1.5 files with "$THREADS" blocks
support for this block should be in the library to be able to keep on
reading those VEX files.