NEWSTRUCT split into declare/alloc variable steps
With the advent of VEX2 there may be cases where a decision is inserted
between the declaration of the new struct-to-create variable and the actual
allocation of memory for that.

To support this cleanly with the compiler warnings cranked up beyond 11(*)
the NEWSTRUCT() canned sequence (declaration + initialization of variable in
one fell swoop) is - behind the scenes - split into separate declaration and initialization
phases, with their own macros.

The NEWSTRUCT() macro is functionally unchanged, but now the developer has
access to the separate NEWSTRUCTDECLARE() and NEWSTRUCTALLOC() macros to be
used as needed.

(*) Spinal Tap reference