Some cleanup/undo deletion of useful stuff/style
Some leftover cleanup:

- useless comments removed

- function prototype mismatch edit changed to solve the issue:
in the original version of the parser get_all_literals() has a prototype
difference between the header and implementation return value specification ("char*" versus "void").
The fix applied was to change both to "void*" and add (!) a bogus "return
0;" statement in the function where there was no return statement at all.
The function cannot ever return anything useful so this fix
undoes the proposed fix and changes the return type to "void" and returns
the get_all_literals() not having a return statement.

- apply vex_parser brace style: opening brace on same line of "if" statement

- The (useful) information content - line number - in an error message that was deleted brought back